pyproject.toml
[tool.poetry]
name = "discordhtml"
version = "0.1.0"
description = "HTML is inside you"
authors = ["elon_duck"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^1.7.3"
pyquery = "^1.4.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
